ADM-436 combine deploy info with github commits

+++ backend/src/main/java/heartbeat/controller/report/dto/response/DeploymentFrequencyModel.java
@@ -18,7 +18,7 @@ public class DeploymentFrequencyModel {
 
 	private String step;
 
-	private double value;
+	private float value;
 
 	private List<DeployInfo> passed;
